Updated Board constructor and corrected error in Difficulty

+++ src/main/java/model/Difficulty.java
@@ -24,7 +24,7 @@ public enum Difficulty {
         return width;
     }
 
-    public long getHeight() {
+    public int getHeight() {
         return height;
     }
 }
